+++ devspaces-idea/get-sources.sh
@@ -38,7 +38,7 @@ if [[ ${PULL_ASSETS} -eq 1 ]]; then
   log "[INFO] java version:"
   java -version
 
-  ./projector.sh build --prepare --url $idePackagingUrl --tag idea-rhel8
+  ./projector.sh build --prepare --url $idePackagingUrl --tag idea-rhel9
 
   if [[ ! -f "asset-ide-packaging.tar.gz" ]]; then
     log "[ERROR] 'asset-ide-packaging.tar.gz' not found, so nothing to build."
@@ -86,7 +86,7 @@ if [[ $(git diff-index HEAD --) ]] || [[ ${PULL_ASSETS} -eq 1 ]]; then
 		git status || true
 		gitbranch="$(git rev-parse --abbrev-ref HEAD)"
 		target=${gitbranch}-containers-candidate
-		if [[ $scratchFlag == "--scratch" ]]; then target=devspaces-3-rhel-8-containers-candidate; fi
+		if [[ $scratchFlag == "--scratch" ]]; then target=devspaces-3-rhel-9-containers-candidate; fi
 		repo="$(git remote -v | grep origin | head -1 | sed -r -e "s#.+/containers/(.+) \(fetch.+#\1#")"
 		sha="$(git rev-parse HEAD)"
 		tmpfile=$(mktemp) && brew container-build ${target} git+https://pkgs.devel.redhat.com/git/containers/${repo}#${sha} --git-branch ${gitbranch} ${scratchFlag} --nowait 2>/dev/null | tee 2>&1 "${tmpfile}"
@@ -105,7 +105,7 @@ else
 		git status || true
 		gitbranch="$(git rev-parse --abbrev-ref HEAD)"
 		target=${gitbranch}-containers-candidate
-		if [[ $scratchFlag == "--scratch" ]]; then target=devspaces-3-rhel-8-containers-candidate; fi
+		if [[ $scratchFlag == "--scratch" ]]; then target=devspaces-3-rhel-9-containers-candidate; fi
 		repo="$(git remote -v | grep origin | head -1 | sed -r -e "s#.+/containers/(.+) \(fetch.+#\1#")"
 		sha="$(git rev-parse HEAD)"
 		tmpfile=$(mktemp) && brew container-build ${target} git+https://pkgs.devel.redhat.com/git/containers/${repo}#${sha} --git-branch ${gitbranch} ${scratchFlag} --nowait 2>/dev/null | tee 2>&1 "${tmpfile}"